Safari-611.1.10/LayoutTests/svg/custom/svg-parse-overflow-3-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
